introduce `channel-margin-ppm`
Working with higher econ ratios near 1 and lost profits can lead
to a situation where the rebalancing costs are almost equal to the
potential gains, especially if the ppm on the source channel is
very low. Using lower ratios is not an alternative if you want
to rebalance between sinks as well.

The new parameter `channel-margin-ppm` adds an additional margin
to the existing margin of (1-ratio)*amount*fee_rate. The additional
addon is determined by the difference between `channel-margin-ppm`
and the fee rate on the source channel.